requirements.txt
python-dotenv
discord.py
google-generativeai
openai